Move decimal logic
