+++ arrow/src/datatypes/mod.rs
@@ -30,7 +30,9 @@ mod numeric;
 pub use numeric::*;
 mod types;
 pub use types::*;
+mod decimal;
 mod delta;
+pub use decimal::*;
 
 pub use arrow_schema::{DataType, Field, IntervalUnit, Schema, TimeUnit, UnionMode};
