+++ arrow/src/datatypes/types.rs
@@ -17,7 +17,7 @@
 
 use super::{ArrowPrimitiveType, DataType, IntervalUnit, TimeUnit};
 use crate::datatypes::delta::shift_months;
-use crate::util::decimal::{
+use crate::datatypes::{
     DECIMAL128_MAX_PRECISION, DECIMAL128_MAX_SCALE, DECIMAL256_MAX_PRECISION,
     DECIMAL256_MAX_SCALE, DECIMAL_DEFAULT_SCALE,
 };
